disagreement_ai_service/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
openai
crewai
langchain-openai
